Move lifecycle tasks into lifecycle plugin
