+++ buildSrc/subprojects/plugins/plugins.gradle.kts
@@ -21,8 +21,8 @@ dependencies {
 gradlePlugin {
     plugins {
         register("buildTypes") {
-            id = "gradlebuild.build-types"
-            implementationClass = "org.gradle.plugins.buildtypes.BuildTypesPlugin"
+            id = "gradlebuild.lifecycle"
+            implementationClass = "org.gradle.plugins.lifecycle.LifecyclePlugin"
         }
         register("buildVersion") {
             id = "gradlebuild.build-version"
